Move create_pipeline_generator() to model utils
